+++ src/components/atoms/Loading/index.stories.tsx
@@ -7,7 +7,6 @@ const meta: Meta<typeof Loading> = {
   parameters: {
     layout: 'centered',
   },
-  tags: ['autodocs'],
 };
 
 export default meta;
