+++ src/components/features/lobby/ParticipantsList/index.stories.tsx
@@ -8,7 +8,6 @@ const meta: Meta<typeof ParticipantsList> = {
   parameters: {
     layout: 'centered',
   },
-  tags: ['autodocs'],
 };
 
 export default meta;
